backend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
anthropic
SpeechRecognition
pyaudio
pydub
gtts